LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_measure_exceptions-expected.txt

Description

This test validates that the performance.measure() method throws a SYNTAX_ERR exception whenever a non-existent mark is provided as the startMark or endMark, and the method also throws a INVALID_ACCESS_ERR whenever a navigation timing attribute with a value of zero is provided as the startMark or endMark.


PASS window.performance is defined 
PASS window.performance.measure("measure", "mark"), where "mark" is a non-existent mark,  threw an exception. 
PASS window.performance.measure("measure", "mark"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "mark", "responseEnd"), where "mark" is a non-existent mark, threw an exception. 
PASS window.performance.measure("measure", "mark", "responseEnd"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "navigationStart", "mark"), where "mark" is a non-existent mark, threw an exception. 
PASS window.performance.measure("measure", "navigationStart", "mark"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "mark", "mark"), where "mark" is a non-existent mark, threw an exception. 
PASS window.performance.measure("measure", "mark", "mark"), where "mark" is a non-existent mark, threw a SYNTAX_ERR exception. 
PASS window.performance.measure("measure", "loadEventEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "loadEventEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
PASS window.performance.measure("measure", "loadEventEnd", "responseEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "loadEventEnd", "responseEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
PASS window.performance.measure("measure", "navigationStart", "loadEventEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "navigationStart", "loadEventEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
PASS window.performance.measure("measure", "loadEventEnd", "loadEventEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an exception. 
PASS window.performance.measure("measure", "loadEventEnd", "loadEventEnd"), where "loadEventEnd" is a navigation timing attribute with a value of 0, threw an INVALID_ACCESS_ERR exception. 
